Disable ligatures in X11 fonts
So I only noticed this in Firefox, but it was a fun one to track down.
Basically, any character pairing that could use a ligature (such as
`ff`, `fi`, etc) and was in the system default font (I believe it was
Cantarell?) would fuck up and use some random unicode symbol (usually a
right arrow of some sort). This made URLs super confusing to see, and
when typing them it made me question my sanity every time (hit `f`, `i`,
and oh shit what now I've typed an arrow)...

Updating the font configuration to not use embedded bitmaps (which I
assume it how the ligatures work) solves the issue.