Hide errors from __git_ps1
Was running into an issue where `sudo su` would fails to load the
sourced git extensions, and thus causes an error on every command line,
which is very noisy and annoying. We can solve this by just piping all
errors from __git_ps1 into oblivion.

This is not "technically" correct, probably, but it works™.