Clean up dexec, move into script
I recently (finally) got rid of the need to use `-lc` with bash to get
the PATH modifications in the containers, so we can just pass the
command right through. This is _great_ because it means I can get rid of
those nasty nested conditionals.

Also since this is such a heavy function, it felt more "correct" to move
this into an actual script in my local bin directory so it's a small
shell script instead of a large function.