Remove old dotfiles
This is a big one. This changeset clears out all the non-standard
platform dotfiles I had lying around "just in case". These are configs
for software and platforms I used briefly but haven't come to use
regularly or, in most cases, haven't used more than once.

In the case of gtk-3.0/settings.ini, I don't even know why that file was
in there. it doesn't do anything and was setting themes and fonts I
haven't seen or installed in years.