Switch to a custom, more muted solarized
Solarized is a fantastic colourscheme, but the constant blue becomes
oppressive after a while, and usually is what drives me to something
else. This modifies the colour palette to drop the brightness and
saturation of most of the colours so they're more "muted", and drops the
background/blacks significantly to make them much more muted and darker.
The end result is really fantastic, I think.