Add fork of Sublime's solarized theme
This more-or-less matches my new customized, toned-down Solarized I made
for the terminal. tmTheme's aren't my cup of XML, so I got as far as
changing the background colour before decided it was good enough for
now. Maybe one day when I care more I'll finish the rest of the palette.