Prevent tmux < 1.9 from maintaining cwd
This is killing me. Using RHEL on my work laptop means tmux 1.8 which
maintains the cwd when you <prefix>c a new window. I've become so
accustomed to this not happening it's frustrating me to no end.

This option _seems_ portable, at least between my work laptop on tmux
1.8 and my personal laptop on 2.5. But tmux has randomly broken options
before so I'm skeptical of this long-term. The alternative is some
absolutely disgusting if-shell hack so yolo.