Fix default-path on newer tmux versions
So default-path doesn't exist in newer tmux versions, but in order to
have consistent behaviour on older versions we need to set it.

Easy right? we just need to check the tmux version! Surely that's
available in the config, right? oh, it's not... Well at least there's an
if/else syntax right????????????? LOL ONLY IN THIS SATANIC FORM!

fuck everything about this. it was painful to write, is painful to look
at, and is painful to know that it's necessary just to have a fucking
consistent and portable experience using this godforsaken software.