Fix vim-elixir's default garbage
They decided on a whim that overwriting `gq`'s default behaviour to call out to
a CLI feature available only in the latest Elixir was a good idea. When asked
about it[1], they just linked to an XKCD (wtf??).

So, because the maintainers wanted to be cute, we have to force it to not
violate my editor by setting the formatprg to empty, which it checks for
before injecting its own.

[1]: https://github.com/elixir-editors/vim-elixir/issues/399