Add a custom version of Solarized Light
This was based off of:

	https://github.com/braver/Solarized

But modified fairly heavily to fix all the various bugs, gross-looking
decisions, and various discrepancies between this version and the
canonical Vim variant.

I think I got it pretty close, to the point where I can now finally
switch Sublime to Solarized Light and not hate every minute of it.

A lot of things I dislike are unfortunately part of the core Sublime
Text Ruby syntax definition, so can't be fixed just in the colourscheme
(things like `Bundler.require` highlighting `require` like the top-level
keyword).

There are many tweaks still to do but this is a reasonable starting off
point.