Add kates-deploy script for k8s deployments
This is a small wrapper around some Docker and kubectl commands to get a
container image built and rolled out to Kubernetes.

I've been using this personally for a couple months and it's been really
handy, so finally comitting it in here.

Requires some conventions to be followed, as in, it requires all
resources to use the same name (generally I use names in the style of
`example-com`)