Add git-co for fuzzy branch checkouts
This adds git-co to replace the old alias for `git co` to pipe a
formatted list of branches to fzy so we can stop guessing branch names
and just either choose or search.

If any arguments are passed, however, it just falls through to
`checkout` normally so things like `git co --` still work as expected
(as well as passing a branch name normally).

I can't believe it took me this long to do this.