Simplify install process and remove multi-platform
This is a big one. We're removing support for macOS from the install
script as I barely am able to use my dotfiles on my work laptop anyway.
This then allows us to significantly simplify the hierarchy and install
script to do away with the "component" folders for each platform and
just have one `configs` folder for everything. This brings `dotfiles`
in-line with `bootstart`, the latter of which lost macOS support a while
back in a similar effort.

We also basically rewrite the whole install script to use `find` more
extensively and overwrite `IFS` instead of weird manual `sed` calls.
Overall, I think I trust this script a bit more than the previous one,
and it's not as embarrassing, so I guess that's a win.

The VSCode extension installation was also moved to its own script since
putting it in `install` was pretty weird and bad.