Remove Spring Boot pack from VSCode
This was nice for Spring development, but was extremely bloated and
started interfering with every YAML file I tried to open.

Not worth it for the very small amount of Spring I deal with. Maybe we
can just install some pieces of it later to pare it down a bit.