Add neovim configuration
This ports over most of my vim configuration to Neovim, with some slight
adjustments -- notably, switching to `fzf` for a fancier fuzzy finder
and using `vim-plug` instead of manual git submodules.

With the move to fzf, we also now set `FZF_DEFAULT_COMMAND`, as the
default `:Files` was including gitignored stuff, which was gross. Using
`ag` makes it behave the same as before where we were manually piping
`ag` into `fzy`.

I was pleasantly surprised how many options I no longer had to set just
to get sane defaults. I hope this is the beginning of a beautiful
friendship.

Also, since I definitely won't be able to get used to typing `nvim`
instead of `vi` or `vim`, I have added `e` as a generic alias to
whatever is set as `$EDITOR`, which is an easier thing to get used to
and ensures if I ever switch again I will never have to relearn.