Minimize zsh prompt
This removes the hostname from the zsh prompt -- I always visually just
ignore it, and I rarely have dotfiles on whatever machine I am ssh'ed
into so it's not really actually helping anything by having it there.

The prompt felt kind of naked without it though, so I changed the git
branch formatting to be in parens instead of dangling, which helped a
bit.

Also changed the prompt itself to a unicode arrow, →, which makes
everything seem a bit lighter as well because of the lower height.