Add a couple more language servers for coc.vim
I write Java and Go both more frequently than I probably was ready to
admit. Both are good languages.

And, as much as I might wish otherwise, I still have an will continue to
have a Ruby job for at least the mid-term future, so might as well
formalize that for coc.