Set cursor colour in Terminal.app
Finally got tired of not being able to read things under the cursor
since it was the same colour as text. This changes it to a 30% opacity
magenta, which is enough contrast to read almost everything under it,
plus is bright and stands out enough so you don't lose it as much.