Add dark variant of Solarized Muted
Sometimes at night I want a dark theme, but would like it to be somewhat
consistent with my de-saturated Solarized theme. I finally just broke
down and made a "solarzed dark muted" to match "solarized muted" which
is solarized muted but with the background inverted, some tweaks to make
the black/brightblack look nicer, and a darker cursor.

It looks pretty badass, ngl