Remove coc-pairs
I liked this in theory but oh man it is a hinderance wayyyy too often,
leaving orphan "pair" closing brackets that you then have to clean up. I
think the amount of time I spent cleaning up orphan paired brackets is
more than it would've taken for me to just type the damn closing
brackets, so unfortunately it is goodbye to coc-pairs.