Set up true colours in vim and tmux
This configures tmux and neovim to use "true colours" so that they can
be themed more specifically and without being affected by the terminal
theme.

By doing so, we can now use something like the github-light vim theme,
which needs very specific colours, and is what I was trying to go for
with my Solarized Muted theme to begin with. This feels much cleaner and
less prone to terminal- or os-specific quirks.