Enable italic support in tmux
This required a bit of digging and a bit of trial-and-error. It seems
like people get very confused about whether or not you need to compile a
custom terminfo profile, when really all you need is to change the
default one in tmux's config...

I also cleaned up the terminal-overrides as the cursor switching appears
to also be fixed automatically by this so we don't need to handle it
ourselves.

The only thing that broke from this was one of the weird up-arrow binds
that referenced some key directly in $terminfo, so since everything
works when I remove that still, I am just going to make peace with that.