Improve tmux aesthetic, add session details
This changes the tmux status-bg to be slightly darker than the default
terminal background to separate it better. Also, left-aligns the windows
and adds the machine hostname and session name to the status-right. This
makes the whole thing more symmetrical, and looks a bit nicer. Plus now
I can start naming my sessions and not be confused which one is 0 and 1.