Show hidden files in fzf
This has bugged me for a while and I'm not sure why I never fixed it
until now: I often want to fzf-find hidden files and can't. This updates
the FZF_DEFAULT_COMMAND to include hidden files, and adds an .fdignore
so that doesn't include all the crazy stuff in .git.