Switch neovim configuration to Lua
Neovim has clearly put their weight behind moving configuration to Lua.
As of 0.5.0 we can now configure everything via 'init.lua', meaning we
have virtually no vimscript involved, even from the very start.

This took a lot of research, and I had to rewrite a bit of my custom
"open terminal window" function, but overall it was pretty smooth and
straightforward.

To work better in this lua world we're also switching to Packer instead
of Plug, as the lua config is rather nice.

Welcome to the future, I guess.