Simplify whitespace highlighting and fix colours
tokyonight does not use a bg colour for ErrorMsg so for spaces it
basically means they are invisible still. Setting the highlight group to
TSWarning highlights them as if they are like a lint warning or
whatever, which is probably accurate, though does limit us to treesitter
themes, which is probably fine.

This also moves this to a multiline string syntax so we don't have to
escape everything, which was causing issues.