Make zsh prompt more minimal
This switches back to the "@branch" format to make it more compact,
switches to an acute single quote bracket for the prompt character, and
changes the branch name colour to cyan which makes it still stand out
beside the blue path but much less so.

This makes the prompt look a lot lighter and harmonious.