Document and allow cancellation of `eg`
Currently if you `^D` out of `eg` it will just open a new editor
instance with no file. This is not intended. We should check if a
selection was made before invoking the editor -- similar to what we
already do in `cdp` for `cd`.

Also, we should update the function to be documented so it's more
helpful and matches the rest of the file.