Refine file opening mechanics in nvim
After much time with \f and \t, I still can't seem to get used to it.
For whatever reason, I constantly mix up f and t and end up opening
nerdtree and rapidly typing a filename and then causing a mess and
ruckus. This will not do.

I'm stealing a bit from the intellij playbook to shake things up -- fzf
is now bound to \\ (leader-leader), similar to intellij's shift-shift to
open their fuzzy finder thing. It's very ergonomic and I quite like it.

For nerdtree, we're moving it to open with \f (for "File" or "Find",
idk) and then \F to locate the current buffer in the tree ("File
harder"). This seems to make more sense in my brain.