Add scrupt to automate dark mode enablement
Sometimes it's dark and I want my computer to match. It's a bit annoying
to have to uncomment a bunch of config whenever I want to switch, so
this adds a CLI to do it for me.

I moved most of the dark mode configs into their own files as finding
the right line to uncomment is easier, especially for kitty where there
are multiple things to change.