Only click links with modifier in kitty
I accidentally click on things that look like links all the time and
after a delay i get torn out of the terminal and the browser tries to
load some broken link-like thing i accidentally clicked on.

To stop this, we unbind left click from the `link` action, and put that
only under the `ctrl+click` modifier, which also means there doesn't
need to be a delay. 1000x better, no idea why it's not the default.