Configure ssh to use 1Password
1Password 8.x introduced the ability to use it as an identity agent with
SSH, which means private keys can be stored securely and never have to
touch the disk.

Not 100% sold on hard-coding this into dotfiles since it's a bit OS- and
machine-specific (eg., I'll need to remove some of this on my work
laptop since I need different SSH keys and hosts). Whatever, it'll be
fine for now and isn't a huge pain to fix.