Switch from packer to git submodules
Packer was kind of annoying in the long run, as it just clones
everything from git anyway, and requires an extra step to set things up.
Plus it itself was a submodule, but also needed to keep itself in its
own package list (??) which meant there was an error for every packer
command saying it failed to maintain itself? But if you removed it from
the list it would delete itself....

Just wasn't worth the hassle compared to just a set of git submodules.