Remove ssh config from dotfiles
This is too machine- and setup-specific to really fit here. I have to
modify this or at the least remove it on any machine or VM that isn't my
main laptop... just not worth the hassle. Will just have this floating
around manually for now, and will try and think of a better place for it
long term.