Remove copilot
This was cool and everything, but now my trial is over and I have to
really decide if I'd pay money for it. And honestly... Probably not.
It's wrong more than it's right, and I'm still a little uneasy about the
privacy implications. Usually it just gives me a pile of code using
deprecated APIs that I have to delete and figure out myself anyway.

Maybe I'll give it another shot one day, especially if there's a
self-hosted version... but not today.