Add generic tool-versions to home directory
In the past this was a pain to keep up to date because of hard-coded
versions, but now rtx lets us use aliases which makes it a lot better,
as globally we do generally just care that it's recent and not what
specific version it is.

Except elixir and erlang, because they need to be kept at compatible
pairings, annoyingly.