Set neovim theme automatically
This sets the neovim colourscheme based on the background setting of
light/dark, which neovim detects based on the terminal's background
colour automatically.

There is sometimes a very brief flash of the wrong scheme at startup
just because we have to do this in an autocmd. Would be nice to find a
way to fix this but it's pretty easy to live with.