Remove Sublime Text config
🫡

Life is a winding road. I will forever love Sublime but I also have to
come to terms that I have not used this for anything other than quick
notes or a scratchpad for the better part of a decade. Definitely not
necessary to keep standardized config hanging around for that.