Add Ghostty config
First impressions are very good for this, especially for a brand new
project. Might just steal me away from Kitty, we'll see.

The no-config philosophy is pretty great so far; the only thing I've
wanted to change is the font to my obscure favourite. Might eventually
want to switch back to a light theme but we'll stick with the default
for now and see how we get on.